requirements.txt
pandas>=2.0
numpy>=1.24
xgboost>=2.0
scikit-learn>=1.3
biopython>=1.81
requests>=2.28
joblib>=1.3
matplotlib>=3.7
